feat: split rspack build and dev test
